src/hiro_graph_client/requirements.txt
requests~=2.25.1
backoff~=1.10.0


setuptools~=56.1.0
websocket-client~=0.59.0
uuid~=1.30